finish cpp solve
