+++ src/cpp/CMakeLists.txt
@@ -3,5 +3,11 @@ cmake_minimum_required(VERSION 3.10)
 project(solver)
 
 add_executable(solver
-    solve.cpp
+    main.cpp
+    node.cpp
+    node.h
+    puzzle.cpp
+    puzzle.h
+    state.cpp
+    state.h
 )
