+++ src/cpp/puzzle.cpp
@@ -1,0 +1,10 @@
+#include "puzzle.h"
+#include "node.h"
+
+Puzzle::Puzzle() : initial_node(nullptr)
+{
+}
+
+Puzzle::~Puzzle()
+{
+}
